experimenting/prototype/requirements.txt
kafka-python
flask
arrow
mypy-lang